Unused logic from refactored ItemPage moved to helpers
